requirements.txt
scamp
pyFluidSynth
opencv-python
imutils
python-rtmidi
pynput
pillow